# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Movie Web App API
  description: |
    REST API for Movie Web App.

    ## Authentication
    Using **Google OAuth2** for login and **JWT Bearer Token** for accessing endpoints that require authentication.

    ## Roles
    - **guest** — Public access without authentication
    - **user** — User that has logged in via Google OAuth
    - **admin** — Administrator with full access
  version: 1.0.0
  contact:
    name: API Support
servers:
  - url: http://localhost:5000
    description: Local development
  - url: https://movie-ku-api.vercel.app
    description: Production
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token obtained from Google OAuth login
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
      required: [success, message]
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
        error:
          type: string
      required: [success, message]

    # --- Auth ---
...
    Genre:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
      required: [id, name]

    # --- Movie Image ---
    MovieImage:
      type: object
      properties:
        id:
          type: string
          format: uuid
        image_type:
          type: string
          enum: [poster, backdrop]
        image_url:
          type: string
        width:
          type: integer
          nullable: true
        height:
          type: integer
          nullable: true

    # --- Movie Video ---
    MovieVideo:
      type: object
      properties:
        id:
          type: string
          format: uuid
        video_type:
          type: string
          enum: [trailer, teaser]
        site:
          type: string
          enum: [youtube, vimeo]
        video_key:
          type: string
        official:
          type: boolean

    # --- Movie ---
    Movie:
      type: object
      properties:
        id:
          type: string
          format: uuid
        api_id:
          type: string
          nullable: true
          description: TMDB API ID (null if user-created)
        source:
          type: string
          enum: [tmdb, user]
        title:
          type: string
        overview:
          type: string
          nullable: true
        release_date:
          type: string
          format: date
          nullable: true
        popularity:
          type: number
          format: float
          nullable: true
        rating:
          type: number
          format: float
          nullable: true
        is_featured:
          type: boolean
        status:
          type: string
          enum: [active, archived]
        created_by:
          type: string
          format: uuid
          nullable: true
        genres:
          type: array
          items:
            $ref: "#/components/schemas/Genre"
        images:
          type: array
          items:
            $ref: "#/components/schemas/MovieImage"
        videos:
          type: array
          items:
            $ref: "#/components/schemas/MovieVideo"
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      required: [id, source, title, status, created_at, updated_at]
...
    Wishlist:
      type: object
      properties:
        id:
          type: string
          format: uuid
        movie_id:
          type: string
          format: uuid
        movie:
          $ref: "#/components/schemas/Movie"
        scheduled_watch_date:
          type: string
          format: date
          nullable: true
        reminder_sent:
          type: boolean
        created_at:
          type: string
          format: date-time
      required: [id, movie_id, created_at]
    WishlistCreateRequest:
      type: object
      properties:
        movie_id:
          type: string
          format: uuid
        scheduled_watch_date:
          type: string
          format: date
      required: [movie_id]
    WishlistUpdateRequest:
      type: object
      properties:
        scheduled_watch_date:
          type: string
          format: date

    # --- Admin Dashboard ---
...
    PaginationMeta:
      type: object
      properties:
        page:
          type: integer
          example: 1
        per_page:
          type: integer
          example: 20
        total:
          type: integer
        total_pages:
          type: integer
  parameters:
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        default: 1
        minimum: 1
      description: Page number
    PerPageParam:
      name: per_page
      in: query
      schema:
        type: integer
        default: 20
        minimum: 1
        maximum: 100
      description: Items per page
...
paths:
...
  /api/wishlists:
    get:
      tags: [Wishlist]
      summary: List user's wishlist
      operationId: listWishlists
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/PerPageParam"
      responses:
        "200":
          description: Wishlist items
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessResponse"
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: "#/components/schemas/Wishlist"
                      meta:
                        $ref: "#/components/schemas/PaginationMeta"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    post:
      tags: [Wishlist]
      summary: Add movie to wishlist
      operationId: createWishlist
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WishlistCreateRequest"
      responses:
        "201":
          description: Added to wishlist
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessResponse"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/Wishlist"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Movie already in wishlist
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "422":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/wishlists/{id}:
    put:
      tags: [Wishlist]
      summary: Update wishlist item
      operationId: updateWishlist
      description: Update scheduled watch date for a wishlist item.
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WishlistUpdateRequest"
      responses:
        "200":
          description: Wishlist updated
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessResponse"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/Wishlist"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Wishlist item not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    delete:
      tags: [Wishlist]
      summary: Remove from wishlist
      operationId: deleteWishlist
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Removed from wishlist
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Wishlist item not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ==================== GENRES ====================
...